Fix table view unsymmetric views & other small fixes
Tables did not allow different amounts of rows compared to headers.
Table and Item context dicts are now slit to required and not required
parts, so that passing None is not required for optional arguments.
Mention in docs about kwargs on the views in docs if admin kwargs are
given.